requirements.txt
click
pygogo
colorama
halo
PyYAML
yamlordereddictloader
PTable
pylint